requirements.txt
asgiref==3.2.10
Babel==2.5.3
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.17
django-bootstrap3==11.0.0
django-comments-dab==2.5.1
django-comments-xtd==2.8.1
django-contrib-comments==1.9.2
django-cors-headers==3.5.0
django-countries==6.1.3
django-email-verification==0.0.2
django-enumfields==2.0.0
django-extensions==3.1.1
django-filer==1.7.1
django-filter==2.2.0
django-import-export==2.3.0
django-jinja==2.5.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-parler==2.0.1
django-polymorphic==2.1.2
django-registration-redux==2.7
django-reversion==3.0.5
django-simple-gravatar==1.0.1
django-timezone-field==3.1
djangorestframework==3.11.0
docutils==0.16
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
factory-boy==2.7.0
fake-factory==0.5.3
gunicorn==20.0.4
haversine==2.2.0
httpie==2.2.0
idna==2.10
jdcal==1.4.1
jeepney==0.6.0
Jinja2==2.8.1
jsonfield==2.1.1
keyring==21.8.0
keyrings.alt==4.0.2
lxml==4.6.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==2.5.14
packaging==20.8
Pillow==7.2.0
psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
SecretStorage==3.3.0
shuup==2.3.9
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.2
unicodecsv==0.14.1
Unidecode==1.1.2
urllib3==1.25.10
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
